+++ Sources/SuperwallKit/StoreKit/Coordinator/CoordinatorProtocols.swift
@@ -36,5 +36,5 @@ protocol TransactionRestorer: AnyObject {
   /// - Returns: A boolean indicating whether the restore request succeeded or failed.
   /// This doesn't mean that the user is now subscribed, just that there were no errors
   /// obtaining the restored transactions
-  func restorePurchases() async -> Bool
+  func restorePurchases() async -> RestorationResult
 }